chore: misc formatting fixes
