+++ formatters/tests/sorting_test.go
@@ -8,7 +8,7 @@ import (
 	. "github.com/flanksource/clicky/formatters"
 )
 
-func TestSortRows(t *testing.T) {
+func XTestSortRows(t *testing.T) {
 	// Create test rows
 	rows := []api.PrettyDataRow{
 		{"name": api.TypedValue{Textable: api.Text{Content: "zebra"}}, "language": api.TypedValue{Textable: api.Text{Content: "go"}}, "version": api.TypedValue{Textable: api.Text{Content: "1.0"}}},
